requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
joblib==1.4.2
python-docx==1.1.2
openpyxl==3.1.5
toml==0.10.2
